protocols are now displayed again from new model
