+++ webcentral/src/protocols/admin.py
@@ -1,3 +1,10 @@
 from django.contrib import admin
+from modeltranslation.admin import TranslationAdmin
 
-# Register your models here.
+from .models import Protocol
+
+class ProtocolAdmin(TranslationAdmin):
+    pass
+
+
+admin.site.register(Protocol, ProtocolAdmin)
